+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * Default key binding to trigger group start.
 
+### Fixed
+
+* Use the current file editor for scrolling down on processes.
+* All file editors viewing a process will follow down if the last row was selected.
+
 ## [0.0.9]
 
 ### Added